requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
openpyxl>=3.1.0
xgboost>=1.7.0
prophet>=1.1.5
pmdarima>=2.0.4
pyecharts>=2.0.6
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
orjson>=3.9.10
pydantic-settings>=2.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.1

